refactor(subscriptions): group all components into per-parent folders
