+++ packages/client/src/pages/Subscriptions/components/CandidatesBanner/CandidateRow.tsx
