+++ packages/client/src/pages/Subscriptions/components/LinkTransactionsModal/index.tsx
@@ -117,12 +117,12 @@ const LinkTransactionsModal = ({ subscription, onClose, onLinked }: Props) => {
                               {format.euro(tx.amount)}
                             </Typography>
                           </Box>
-                      }
+                        }
                         secondary={
                           <Typography variant='caption' color='textSecondary'>
                             {tx.category?.name} · {tx.account?.name}
                           </Typography>
-                      }
+                        }
                       />
                     </ListItem>
                   )
